requirements.txt

python-dotenv
snowflake-sqlalchemy
sqlalchemy
pandas
psycopg2-binary
